Fixed or tried to fix the last little things.
The things that could be fixed are fixed. The API calls that run empty in the backend were skipped and will be implemented in the next sprint instead.
The forgotten password API call could not be fixed.

All and remaining sub-task (MH-XX) by Josina and me. The Changes are independent of this and have nothing to do with the co-author.
MH-20 MH-22 MH-26
MH-30 MH-31 MH-32 MH-35 MH-36 MH-37 MH-38 MH-39
MH-43 MH-44 MH-45 MH-48 MH-49
MH-50 MH-51 MH-52 MH-53 MH-54

Co-Authored-By: josina1810 <95410472+josina1810@users.noreply.github.com>